Distinguish between verified and unverified blocks (#114)
When we receive a block from the network, it should be verified before it can be persisted into stable storage.

Similarly, when we load a block from stable storage, there is no more need to verify it, as we had verified it before.

This commit makes the compiler enforce that a block received from the network cannot be persisted into stable storage before it has been verified.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>