Make sure QC only contains authorized signers and is not signed twice (#118)
This commit makes sure that a QC only contains signers that are among the nodes the Epoch was initialized with.
Similarly, it checks the QC is not signed twice or not signed by too few nodes.

This eases the implementation of a QC by the application, especially for naive implementations where the QC contains
the public keys of the signers, and the Signers() method just returns a deterministic derivation of the aforementioned public keys.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>